Extract some utils to ember-shared
- these were duplicated across ember apps
