+++ packages/ember-shared/addon/src/index.ts
@@ -1,1 +1,4 @@
 export { default as truncateMiddle } from './helpers/truncate-middle';
+export { type Emitter, type UnbindEventListener, SimpleEmitter } from './utils/events';
+export { type VoidCallback } from './utils/types';
+export { MockBroadcastChannel, MockLocalStorage } from './utils/browser-mocks';
