+++ packages/ember-shared/addon/src/utils/events.ts
@@ -2,6 +2,7 @@
 export type UnbindEventListener = () => void;
 
 export interface Emitter<EmitterEvent = string> {
+  // eslint-disable-next-line @typescript-eslint/ban-types
   on(event: EmitterEvent, cb: Function): UnbindEventListener;
 }
 
@@ -14,7 +15,7 @@ export interface Emitter<EmitterEvent = string> {
  * import {
  *   SimpleEmitter,
  *   UnbindEventListener,
- * } from '@cardstack/web-client/utils/events';
+ * } from '@cardstack/ember-shared';
  * import Service from '@ember/service';
  *
  * export default class ResizedTarget extends Service {
@@ -59,16 +60,18 @@ export interface Emitter<EmitterEvent = string> {
  */
 
 export class SimpleEmitter {
+  // eslint-disable-next-line @typescript-eslint/ban-types
   events: Record<string, Function[]> = {};
 
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   emit(event: string, ...args: any[]) {
     (this.events[event] || []).forEach((i) => i(...args));
   }
 
+  // eslint-disable-next-line @typescript-eslint/ban-types
   on(event: string, cb: Function): UnbindEventListener {
     (this.events[event] = this.events[event] || []).push(cb);
-    let unbind = () =>
-      (this.events[event] = (this.events[event] || []).filter((i) => i !== cb));
+    const unbind = () => (this.events[event] = (this.events[event] || []).filter((i) => i !== cb));
     return unbind;
   }
 }
