+++ packages/web-client/app/utils/web3-strategies/layer2-chain.ts
@@ -8,7 +8,11 @@ import { IConnector } from '@walletconnect/types';
 import WalletConnectProvider from '@cardstack/wc-provider';
 import { task } from 'ember-concurrency-decorators';
 
-import { Emitter, SimpleEmitter, UnbindEventListener } from '../events';
+import {
+  type Emitter,
+  SimpleEmitter,
+  type UnbindEventListener,
+} from '@cardstack/ember-shared';
 import {
   ConvertibleSymbol,
   ConversionFunction,
