+++ packages/web-client/app/utils/web3-strategies/test-layer1.ts
@@ -15,10 +15,7 @@ import BN from 'bn.js';
 import { WalletProvider } from '../wallet-providers';
 import { TransactionReceipt } from 'web3-core';
 import { BridgeValidationResult } from '@cardstack/cardpay-sdk';
-import {
-  SimpleEmitter,
-  UnbindEventListener,
-} from '@cardstack/web-client/utils/events';
+import { SimpleEmitter, UnbindEventListener } from '@cardstack/ember-shared';
 import { fromWei, toWei } from 'web3-utils';
 import {
   BridgeableSymbol,
